Change build state logic
It still follows the Arne Blankerts definition but was modified because
I found some issues like interfaces and empty clases being marked as
untested when they should be marked as empty.

Additionally I found some cases where test covered classes were marked
as empty because the executable lines was equal to 0 (not sure if it is
a PHP_CodeCoverage issue or a phpDox issue).

The change in the XSLT logic solves the issues above. I basically
inverted the logic statements. If there is coverage or executable lines
for the class, it will classify it between untested, failed or passed.
Otherwise, it will be classified as empty.

See: https://github.com/theseer/phpdox/issues/147#issuecomment-57431624